Preserve case in words.
In the ANAInfo class, when an analysis is added check the root and
determine which case format it is in and set an internal value that
corresponds to the \c marker. Also always make the root lower case.
Added methods for setting and getting the capitalization number.
setAnalysisByPart no longer takes a list of prefixes and suffixes,
rather a string. Reduce blank spaces in the ANA file. I replaced places
where the ANA records were being written manually to a file and now use
the ANAInfo.write() method.