Version 4 - Preserve case in words.
Do capitalization of the extracted word. Use the baseline word as a
model for how to capitalize it. For the complex form case, save the 1st
base-line word for a model of the whole complex form. Removed large
commented section of code that was for checking if the whole text had
been analyzed.