Major overhaul.
Use class objects to model the link information. Allow the user to link
any sense in the text not just the unlinked ones matched a gloss as
before. Also use a fuzzy compare logic for near matches on gloss to give
more suggestions. Now a list of target senses is provided which can be
used as the sense to link to for any source sense. Also, handle variants
of senses.