Support advanced apertium processing and font size changes
The tool now supports advanced transfer processing. When the .t2x and
.t3x files are present, advanced mode is enabled and the Interchunk and
Postchunk tabs can be used. The output from one gets copied to the input
of the other when the tab is changed. Don't prompt the user for the
transfer rules file anymore, just use the one defined in the
configuration file by default. Copy the bilingual dictionary file and
the Makefile once upon startup to the LiveRuleTester folder. If the
biling. dictionary changes. The tool will have to be restarted or you
can browse to find the file again. Removed unused RuleList class.