complex form bug fix
Fixed bug where two entries that occur together but were not part of the
same complex form cause the 2nd one to be ignored. E.g. dust enteqad kon
the code saw dust was part of a complex form, when it got to enteqad it saw
it too was part of a complex form, but it wasn't in the right position so it
ignored it and the complex form enteqad kon didn't get recognized. Now when
enteqad is reached, the counter starts over for looking for new complex forms.