Rewrite of the MainFunction.
Use the Utils.getInterlinData function to get
the words in the interlinear text. The advantage is that the Utils function
puts together the phrasal verbs and the like. Also modularized the MainFunction.
Also added the feature of filtering the list to not show proper nouns. Also
show X.X form of the headword, but only if it isn't 1.1. Also only don't do
the fuzzy compare for proper nouns. Also trimmed down the url link to the
target DB. It matches what FLEx 9.1 uses. Fixed bug where the 2nd time an
entry with exact or fuzzy match wasn't getting all the matching rows that
were there the first time. Now I save the matching rows and add them again.