Fixes #6 where only the test results -50 were being displayed
This resulted in having the a test results file with only 1 result to not show anything. Now there's a MAX_RESULTS_TO_DISPLAY set to 50 and the tool will display up to this amount of test results.
Also performance tweaks. the main one that helped is just creating the QTGui object for the green check, red x and yellow triangle just once for a TestResultItem.