fixes #236 added close and apply/close buttons
Also added logic to detect if a check combo box was changed. This is so the user can be prompted to save.
Also made the new buttons smaller and to the right.
Do less prompting when saving. Only say 'saved' on the Apply button. Otherwise when things were saved, put them in the report.