Fixes #399. Also fix tooltip font size
Show **none** on the left size when the source language is in a RTL script.
Also make the tooltip size for all widgets the same as other widgets according to how much the user has zoomed in or out.
Use the unicode literal string for the arrow strings instead of a \u value in order to self-document.