#1148 Sense link setting name, Standardize Custom Field functions
These functions work for getting any custom field at the specified levels.  They are needed in multiple places in the code.  Adjusting some of the names so they don't look like they only apply to one setting. And fixed a bug that caused the wrong value to show, if (none) is an option.

Also rename setting for custom sense link.  (I did not change the name of the "SourceCustomFieldForEntryLink" setting to refer to "Sense", because that would involve migrating users' .config file.)